% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{re-exports}
\alias{re-exports}
\alias{\%>\%}
\title{Re-exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from this package to avoid listing the packages used
as Depends
}
