% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uBands.R
\name{uBands}
\alias{uBands}
\title{Draw bands between series}
\usage{
uBands(uplot, lower, upper, fill, ...)
}
\arguments{
\item{uplot}{Chart created with \code{\link[=uPlot]{uPlot()}}.}

\item{lower}{Name of the lower serie.}

\item{upper}{Name of the upper serie.}

\item{fill}{Fill color.}

\item{...}{Other options.}
}
\value{
An \code{htmlwidget} object of class \code{"uPlot"}.
}
\description{
Draw bands between series
}
