% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSeries.R
\name{uColors}
\alias{uColors}
\title{Series colors}
\usage{
uColors(uplot, ...)
}
\arguments{
\item{uplot}{Chart created with \code{\link[=uPlot]{uPlot()}}.}

\item{...}{Colors to attribute to each series.}
}
\value{
An \code{htmlwidget} object of class \code{"uPlot"}.
}
\description{
Allow to specify series colors in one call.
}
\examples{
uPlot(temperatures[, c("date", "temperature")]) \%>\%
  uColors(temperature = "black")

uPlot(temperatures[, c("date", "low", "high")]) \%>\%
  uColors(
    low = "blue",
    high = "red"
  )
}
