% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uCursor.R
\name{uCursor}
\alias{uCursor}
\title{Cursor configuration}
\usage{
uCursor(uplot, ...)
}
\arguments{
\item{uplot}{Chart created with \code{\link[=uPlot]{uPlot()}}.}

\item{...}{Cursor options, see examples.}
}
\value{
An \code{htmlwidget} object of class \code{"uPlot"}.
}
\description{
Cursor configuration
}
\examples{
library(uPlot)

# disable crosshairs
uPlot(eco2mix[, c(1, 2)]) \%>\%
  uCursor(x = FALSE, y = FALSE)

# zoom on y only
uPlot(eco2mix[, c(1, 2)]) \%>\%
  uCursor(
    drag = list(y = TRUE, x = FALSE)
  )
}
