% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uOptions.R
\name{uOptions}
\alias{uOptions}
\title{Options for uPlot}
\usage{
uOptions(uplot, ...)
}
\arguments{
\item{uplot}{Chart created with \code{\link[=uPlot]{uPlot()}}.}

\item{...}{Options for the chart.}
}
\value{
An \code{htmlwidget} object of class \code{"uPlot"}.
}
\description{
Options for uPlot
}
\examples{
uPlot(eco2mix[1:24, c(1, 2)]) \%>\%
  uOptions(
    title = "Title for the chart"
  )
}
