% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{uProxy}
\alias{uProxy}
\title{Proxy for uPlot}
\usage{
uProxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shinyId}{single-element character vector indicating the output ID of the
chart to modify (if invoked from a Shiny module, the namespace will be added
automatically).}

\item{session}{the Shiny session object to which the chart belongs; usually the
default value will suffice.}
}
\value{
A \code{uProxy} object.
}
\description{
Proxy for uPlot
}
\examples{
\dontrun{

# Consider having created a editor widget with
uPlotOutput("mychart") # UI
output$mychart <- renderUPlot({}) # Server

# Then you can call proxy methods in observer:


}
}
