% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{uRedraw}
\alias{uRedraw}
\title{Redraw a chart}
\usage{
uRedraw(uplot, rebuildPaths = TRUE, recalcAxes = TRUE)
}
\arguments{
\item{uplot}{Proxy for a chart created with \code{\link[=uProxy]{uProxy()}}.}

\item{rebuildPaths}{If \code{FALSE} use cached series.}

\item{recalcAxes}{Recalculate axes.}
}
\value{
An \code{htmlwidget} object of class \code{"uPlot"} or a \code{uProxy} object.
}
\description{
Redraw a chart
}
\examples{
# Toggle visibility with Proxy --------------------------------------------

library(shiny)
library(uPlot)

ui <- fluidPage(
  tags$h3("uPlot update series with proxy"),
  fluidRow(
    column(
      width = 4,
      textInput(
        inputId = "label",
        label = "Label:",
        value = "Consumption"
      ),
      selectInput(
        inputId = "stroke",
        label = "Stroke color:",
        choices = c("firebrick", "steelblue", "forestgreen", "goldenrod")
      ),
      actionButton("update", "Update plot")
    ),
    column(
      width = 8,
      uPlotOutput(outputId = "plot", height = "500px")
    )
  )
)

server <- function(input, output, session) {

  output$plot <- renderUPlot({
    uPlot(data = eco2mix[1:1500, c(1, 2)]) \%>\%
      uSeries("consumption", label = "Consumption", stroke = "firebrick")
  })

  observeEvent(input$update, {
    uProxy("plot") \%>\%
      # First delete existing serie
      uDelSeries(seriesIdx = 2) \%>\%
      # Then add with new parameters
      uAddSeries(
        seriesIdx = 2,
        label = input$label,
        stroke = input$stroke
      ) \%>\%
      uRedraw(rebuildPaths = TRUE, recalcAxes = FALSE)
  })

}

if (interactive())
  shinyApp(ui, server)
}
