% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uScatter.R
\name{uScatter}
\alias{uScatter}
\title{Create a scatter chart}
\usage{
uScatter(
  x,
  y,
  color = NULL,
  default_stroke = "black",
  label_x = NULL,
  label_y = NULL,
  options = list(),
  use_gzipped_json = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{x}{Numeric vector to use on x-axis.}

\item{y}{Numeric vector to use on y-axis.}

\item{color}{A vector to identify different data series.}

\item{default_stroke}{Default stroke color to use if \code{color} is \code{NULL}.}

\item{label_x, label_y}{Labels for x and y axes.}

\item{options}{Options to generate the plot.}

\item{use_gzipped_json}{Use \code{\link[jsonlite:gzjson]{jsonlite::as_gzjson_b64()}} to compress JSON data.}

\item{width, height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\value{
An \code{htmlwidget} object of class \code{"uPlot"}.
}
\description{
Create a scatter chart
}
\examples{

library(uPlot)

uScatter(cars$speed, cars$dist)
}
