% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSeries.R
\name{uSeries}
\alias{uSeries}
\title{Series options}
\usage{
uSeries(uplot, name, label = NULL, stroke = NULL, ...)
}
\arguments{
\item{uplot}{Chart created with \code{\link[=uPlot]{uPlot()}}.}

\item{name}{Name of the serie (column's name in data).}

\item{label}{Label to display in legend.}

\item{stroke}{Stroke color of the line.}

\item{...}{Other options.}
}
\value{
An \code{htmlwidget} object of class \code{"uPlot"}.
}
\description{
Series options
}
\examples{

# Change label and color
uPlot(temperatures[, c("date", "temperature")]) \%>\%
  uSeries(
    name = "temperature",
    label = "Temperature in 2022",
    stroke = "steelblue"
  )

# Hide a serie at start
uPlot(temperatures) \%>\%
  uSeries(
    name = "temperature",
    show = FALSE
  )
}
