% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{uSetData}
\alias{uSetData}
\title{Set chart data}
\usage{
uSetData(uplot, data)
}
\arguments{
\item{uplot}{Chart created with \code{\link[=uPlot]{uPlot()}} or \code{\link[=uProxy]{uProxy()}}.}

\item{data}{The \code{data.frame} to use in chart.}
}
\value{
An \code{htmlwidget} object of class \code{"uPlot"} or a \code{uProxy} object.
}
\description{
Set chart data
}
\examples{
uPlot() \%>\%
 uSetData(temperatures)

# setData with Proxy ------------------------------------------------------

library(shiny)
library(uPlot)

ui <- fluidPage(
  tags$h3("uPlot set data with proxy"),
  selectInput(
    inputId = "variable",
    label = "Energy source:",
    choices = names(eco2mix)[-c(1, 2)]
  ),
  uPlotOutput(outputId = "plot", height = "500px")
)

server <- function(input, output, session) {

  output$plot <- renderUPlot({
    uPlot(data = eco2mix[, c("datetime", "fuel")]) \%>\%
      uSeries(serie = "fuel", label = "Production (MW)", stroke = "blue")
  })

  observeEvent(input$variable, {
    uProxy("plot") \%>\%
      uSetData(eco2mix[, c("datetime", input$variable)])
  }, ignoreInit = TRUE)

}

if (interactive())
  shinyApp(ui, server)
}
