% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uPlot.R
\name{uPlot}
\alias{uPlot}
\title{uPlot}
\usage{
uPlot(data, options, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Data to plot as a \code{list} (must match \code{µPlot.js} expectations) or
a \code{data.frame} where first column is the x-axis and the others the series to plot.}

\item{options}{Options to generate the plot.}

\item{...}{Additional arguments.}

\item{width, height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\value{
An \code{htmlwidget} object.
}
\description{
Fast, memory-efficient Canvas 2D-based charts with Javascript library \href{https://github.com/leeoniya/uPlot}{µPlot}.
}
\examples{
uPlot::uPlot(
  data = list(
    as.numeric(eco2mix$date),
    eco2mix$consumption
  ),
  options = list(
    title = "Electricity consumption in France (2012 - 2020)",
    series = list(
      list(label = "Time"),
      list(label = "Consumption", stroke = "#088A08")
    )
  )
)
}
