% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSeries.R
\name{uSeries}
\alias{uSeries}
\title{Series options}
\usage{
uSeries(uplot, serie, label = NULL, stroke = NULL, ...)
}
\arguments{
\item{uplot}{Chart created with \code{\link[=uPlot]{uPlot()}}.}

\item{serie}{Name (column's name in data) or index of the serie.}

\item{label}{Label to display in legend.}

\item{stroke}{Stroke color of the line.}

\item{...}{Other options.}
}
\value{
An \code{htmlwidget} object of class \code{"uPlot"}.
}
\description{
Series options
}
\examples{

# Change label and color
uPlot(temperatures[, c("date", "temperature")]) \%>\%
  uSeries(
    serie = "temperature",
    label = "Temperature in 2022",
    stroke = "steelblue"
  )

# Hide a serie at start
uPlot(temperatures) \%>\%
  uSeries(
    serie = "temperature",
    show = FALSE
  )

### Line style
# default
uPlot(temperatures[11:40, c("date", "temperature")]) \%>\%
  uSeries(
    serie = "temperature",
    paths = JS("uPlot.paths.linear()"),
    points = list(show = TRUE)
  )
# spline curve
uPlot(temperatures[11:40, c("date", "temperature")]) \%>\%
  uSeries(
    serie = "temperature",
    paths = JS("uPlot.paths.spline()"),
    points = list(show = TRUE)
  )
# stepped line
uPlot(temperatures[11:40, c("date", "temperature")]) \%>\%
  uSeries(
    serie = "temperature",
    paths = JS("uPlot.paths.stepped({align: 1})"),
    points = list(show = TRUE)
  )
}
